+++ packages/service/src/plugin/cache/index.ts
@@ -1,7 +1,7 @@
 /*
  * @Author: 邱狮杰
  * @Date: 2023-01-14 12:05:06
- * @LastEditTime: 2023-01-30 13:47:21
+ * @LastEditTime: 2023-03-26 10:11:54
  * @Description:
  * @FilePath: /memo/packages/service/src/plugin/cache/index.ts
  */
@@ -10,13 +10,14 @@ import { beforeTriggerResultTypes, initializeConfigurationTypes, interceptorImpl
 import { AxiosResponse } from 'axios'
 import { defaultCacheRule, hasCacheConfig, requestConfig } from './config'
 import { CacheData, ExpirationTime } from './utils'
+import { colors, HttpLog } from '../logs/utils'
 
 const cacheHandler = new CacheData()
 
 export { ExpirationTime, requestConfig }
 
 export class CachePrerequisites {
-  private config: requestConfig = {}
+  private config: initializeConfigurationTypes = {}
 
   constructor(config: requestConfig) {
     this.config = config
@@ -46,7 +47,7 @@ export class Cache implements interceptorImpl {
    * @description 当符合缓存标准时 把当前响应缓存起来
    */
   responseSuc(response: AxiosResponse) {
-    const config: requestConfig = response.config
+    const config: initializeConfigurationTypes = response.config
     const rule = config?.cacheRules?.(config) || defaultCacheRule(config)
     // 填充缓存
     cacheHandler.fillTheCache(rule, response.data)
@@ -56,7 +57,7 @@ export class Cache implements interceptorImpl {
   /**
    * @description 开启缓存的情况下会 缓存一个 rule string 并且 给当前rule 一个过期的时间戳
    */
-  requestSuc(req: requestConfig) {
+  requestSuc(req: initializeConfigurationTypes) {
     const rule = req.cacheRules?.(req) || defaultCacheRule(req)
 
     if ((req.useCache || !cacheHandler.hasCache(rule)) && req.cacheExpirationTime) {
@@ -69,6 +70,7 @@ export class Cache implements interceptorImpl {
 }
 
 export class CacheTrigger implements triggerInterceptorImpl {
+  displayName: string = 'CacheTrigger'
   beforeTrigger(config: initializeConfigurationTypes): any | beforeTriggerResultTypes<any> {
     const cachePrerequisites = new CachePrerequisites(config || {})
     if (cachePrerequisites.areThereCachePrerequisites()) {
@@ -76,8 +78,20 @@ export class CacheTrigger implements triggerInterceptorImpl {
       // 尝试使用缓存
       const [cacheExists, cache] = cachePrerequisites.useCache()
       if (cacheExists) {
+        const logs = new HttpLog(config)
+        // console.groupCollapsed(`%c triggerInterceptorImpl:CacheTrigger:beforeTrigger`, `color: ${colors.triggerFrontAndRearTnterceptors}`)
+        // console.log('use cache!', config)
+        // console.groupEnd()
         return { directReturnValue: true, data: cache }
       }
     }
+  }
+  logsCallback(type: 'afterTrigger' | 'beforeTrigger', data: void | initializeConfigurationTypes | beforeTriggerResultTypes<unknown>): void {
+    if (type !== 'beforeTrigger') return
+    if (typeof data !== 'object') return
+    if (!Reflect.get(data, 'directReturnValue')) return
+    console.groupCollapsed(`%c successfully used cache!`, 'color: green')
+    console.log(Reflect.get(data, 'data'))
+    console.groupEnd()
   }
 }