+++ packages/service/src/plugin/logs/error.ts
@@ -1,7 +1,7 @@
 /*
  * @Author: 邱狮杰
  * @Date: 2023-03-23 09:39:21
- * @LastEditTime: 2023-03-23 10:08:21
+ * @LastEditTime: 2023-03-27 08:54:22
  * @Description:
  * @FilePath: /memo/packages/service/src/plugin/logs/error.ts
  */
@@ -15,7 +15,8 @@ export class ErrorWithAxios extends AnomalousChain {
 
   constructor(protected obj: object) {
     super()
-    if (!Reflect.get(obj, 'name') && !Reflect.get(obj, 'code')) {
+
+    if (typeof obj !== 'object' || (!Reflect.get(obj, 'name') && !Reflect.get(obj, 'code'))) {
       this.setErrors(Errors.New('not a Axios Errors!'))
     }
   }
@@ -27,7 +28,9 @@ export class ErrorWithAxios extends AnomalousChain {
 
   @panicProcessing()
   AxiosError() {
+    if (typeof this.obj !== 'object') return this
     if (!(this.obj instanceof AxiosError)) return this
+    if (this.obj?.code === 'ERR_CANCELED') return this
     const err = this.obj as AxiosError
     console.groupCollapsed(
       `%c RESPONSE ERROR %c code:${err.code} %c message:${err.message} %c ${err.config?.baseURL ? err.config.baseURL + err.config?.url : err.config?.url} %c 展开查看接口兜底值:👇`,
@@ -38,7 +41,7 @@ export class ErrorWithAxios extends AnomalousChain {
       ''
     )
     console.log(err)
-    console.log('兜底值 =>', (err.config as initializeConfigurationTypes).pocketValue)
+    console.log('兜底值 =>', (err.config as initializeConfigurationTypes)?.pocketValue)
     console.groupEnd()
     return this
   }